chore: split Contacts.ts file and its tests into separate files for every function
