+++ apps/meteor/app/livechat/server/lib/contacts/mergeContacts.ts
@@ -1,0 +1,4 @@
+import type { ILivechatContact } from '@rocket.chat/core-typings';
+import { makeFunction } from '@rocket.chat/patch-injection';
+
+export const mergeContacts = makeFunction(async (_contactId: string, _visitorId: string): Promise<ILivechatContact | null> => null);
