+++ apps/meteor/app/livechat/server/lib/contacts/shouldTriggerVerificationApp.ts
@@ -1,0 +1,6 @@
+import type { ILivechatContact, IOmnichannelSource } from '@rocket.chat/core-typings';
+import { makeFunction } from '@rocket.chat/patch-injection';
+
+export const shouldTriggerVerificationApp = makeFunction(
+	async (_contactId: ILivechatContact['_id'], _source: IOmnichannelSource): Promise<boolean> => false,
+);
