+++ apps/meteor/app/livechat/server/lib/contacts/validateContactManager.ts
@@ -1,0 +1,9 @@
+import type { IUser } from '@rocket.chat/core-typings';
+import { Users } from '@rocket.chat/models';
+
+export async function validateContactManager(contactManagerUserId: string) {
+	const contactManagerUser = await Users.findOneAgentById<Pick<IUser, '_id'>>(contactManagerUserId, { projection: { _id: 1 } });
+	if (!contactManagerUser) {
+		throw new Error('error-contact-manager-not-found');
+	}
+}
